change some variable types

+++ src/core/Entity.cpp
@@ -76,7 +76,7 @@ void Entity::update() {
 std::string Entity::_toString(unsigned short indent) const {
 	std::stringstream ss;
 	auto put_indent = [&ss] (unsigned short indent) -> std::stringstream& { 
-		for (unsigned short i = 0; i < indent; ++i) 
+		for (unsigned i = 0; i < indent; ++i) 
 			ss << "    ";
 		return ss;
 	};
