+++ src/lifish/entities/AxisBullet.cpp
@@ -69,7 +69,7 @@ AxisBullet::AxisBullet(const sf::Vector2f& pos, lif::Direction dir,
 	// 	- if directionality == 1, [motion frames] [destroy frames]
 	//	- if == 2, [up/down frames] [left/right frames] [destroy frames]
 	//	- if == 4, [down] [up] [right] [left] [destroy]
-	for (unsigned short i = 0; i < data.nMotionFrames && i < 8 / data.directionality; ++i)
+	for (unsigned i = 0; i < data.nMotionFrames && i < 8 / data.directionality; ++i)
 		a_move.addFrame(sf::IntRect(
 				(data.nMotionFrames * d + i) * TILE_SIZE,
 				(info.id - 1) * TILE_SIZE,
@@ -77,7 +77,7 @@ AxisBullet::AxisBullet(const sf::Vector2f& pos, lif::Direction dir,
 				TILE_SIZE));
 
 	// destroy animations are non-directional
-	for (unsigned short i = 0; i < data.nDestroyFrames && i < 5; ++i)
+	for (unsigned i = 0; i < data.nDestroyFrames && i < 5; ++i)
 		a_destroy.addFrame(sf::IntRect(
 				(data.nMotionFrames * data.directionality + i) * TILE_SIZE,
 				(info.id - 1) * TILE_SIZE,
