+++ src/lifish/entities/Enemy.cpp
@@ -111,7 +111,7 @@ Enemy::Enemy(const sf::Vector2f& pos, unsigned short id, const lif::EnemyInfo& i
 	auto& a_right = animated->addAnimation("walk_right");
 	auto& a_left = animated->addAnimation("walk_left");
 
-	for (unsigned short i = 0; i < WALK_N_FRAMES; ++i) {
+	for (unsigned i = 0; i < WALK_N_FRAMES; ++i) {
 		a_down.addFrame(sf::IntRect(
 					i * TILE_SIZE, 
 					0,
@@ -163,7 +163,7 @@ Enemy::Enemy(const sf::Vector2f& pos, unsigned short id, const lif::EnemyInfo& i
 				TILE_SIZE));
 
 	auto& a_death = animated->addAnimation("death");
-	for (unsigned short i = 0; i < death_n_frames; ++i) 
+	for (unsigned i = 0; i < death_n_frames; ++i) 
 		a_death.addFrame(sf::IntRect((WALK_N_FRAMES + i) * TILE_SIZE, 2 * TILE_SIZE, TILE_SIZE, TILE_SIZE));
 
 	auto& animatedSprite = animated->getSprite();
