+++ src/lifish/entities/Teleport.cpp
@@ -34,7 +34,7 @@ Teleport::Teleport(const sf::Vector2f& pos)
 	addComponent(new lif::Sounded(*this, { std::make_pair("warp", lif::getAsset("test", "teleport.ogg")) }));
 
 	auto& anim = animated->addAnimation("teleport");
-	for (unsigned short i = 0; i < N_ANIM_FRAMES; ++i)
+	for (unsigned i = 0; i < N_ANIM_FRAMES; ++i)
 		anim.addFrame(sf::IntRect(i * TILE_SIZE, 0, TILE_SIZE, TILE_SIZE));
 	
 	auto& animatedSprite = animated->getSprite();
